Migrate findByIdOrError to the new transaction model
